full code migration from the repository https://github.com/OlesyaLo/goit-js-hw-12-draft was successful
when creating a repository for this project, I forgot to set up the settings for displaying gh-pages. I realised this mistake only after the final upload of the code to the repository. So I made a new repository and moved the code there, and renamed it as https://github.com/OlesyaLo/goit-js-hw-12-draft